petzone_backend/requirements.txt
Django==4.2.7
gunicorn==21.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0